Agregar los nuevos headers y navbar y un poco de refactor
