+++ src/components/Header/Header.css
@@ -1,0 +1,24 @@
+/* Header.css */
+.header {
+    position: relative;
+    height: 70vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    text-align: center;
+    background-size: cover;
+    background-position: center;
+    background-image: url('background.webp'); /* Replace this with the actual background image */
+}
+
+.header-content h1 {
+    font-size: 3rem;
+    margin-bottom: 0.5rem;
+}
+
+.header-content p {
+    font-size: 1.5rem;
+    margin-top: 0;
+}
