+++ src/components/NavBar/NavBar.js
@@ -1,0 +1,18 @@
+// NavBar.js
+import React from 'react';
+import './NavBar.css'; // You can style it as needed
+
+const NavBar = () => {
+    return (
+        <nav className="navbar">
+            <ul className="nav-links">
+                <li><a href="#home">Home</a></li>
+                <li><a href="#about">About</a></li>
+                <li><a href="#services">Services</a></li>
+                <li><a href="#contact">Contact</a></li>
+            </ul>
+        </nav>
+    );
+};
+
+export default NavBar;
