+++ src/helpers/RentOperations.js
